+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+Unreleased
+----------
+
+* Add survey download button and file importer.
+* Add "widget_key" to survey data to help restore survey data for displayed widgets.
+
 0.1.0 (2023-03-23)
 ------------------